refactor(backup): extract restore paths and lock into backup/paths.rs
