+++ src/application/backup/mod.rs
@@ -1,7 +1,6 @@
 use std::path::{Path, PathBuf};
 
 use rusqlite::{Connection, OpenFlags};
-use serde::{Deserialize, Serialize};
 
 use crate::infrastructure::persistence::settings_repo::{
     DEVICE_LOCAL_SETTINGS, DEVICE_LOCAL_SETTING_PREFIXES, SENSITIVE_SETTINGS,
@@ -21,10 +20,16 @@ pub use manifest::{
 };
 
 mod fs_util;
+mod paths;
 mod snapshot_db;
 
 pub use fs_util::{assert_no_sidecars, SIDECAR_SUFFIXES};
 use fs_util::{crc32_of_file, fsync_dir, fsync_file, sidecar_paths};
+use paths::restore_state_parent;
+pub use paths::{
+    activate_restore_lock, import_staging_dir, pending_restore_dir,
+    restore_bak_dir, restore_lock_active, RestoreState, RESTORE_STATE_PATH,
+};
 use snapshot_db::snapshot_device_id;
 pub use snapshot_db::{
     audio_paths_from_snapshot, open_read_only, snapshot_counts,
@@ -255,51 +260,6 @@ pub fn reveal_in_file_manager(path: &Path) {
     {
         let _ = path;
     }
-}
-
-static RESTORE_LOCK: std::sync::atomic::AtomicBool =
-    std::sync::atomic::AtomicBool::new(false);
-
-pub fn activate_restore_lock() {
-    RESTORE_LOCK.store(true, std::sync::atomic::Ordering::SeqCst);
-    eprintln!("[backup] restore lock active: restart required");
-}
-
-pub fn restore_lock_active() -> bool {
-    RESTORE_LOCK.load(std::sync::atomic::Ordering::SeqCst)
-        || pending_restore_dir().join(DB_ENTRY_PATH).exists()
-}
-
-fn restore_state_parent() -> PathBuf {
-    #[cfg(target_os = "ios")]
-    {
-        let home = std::env::var("HOME").expect("HOME not set on iOS");
-        PathBuf::from(home).join("Library/Application Support")
-    }
-    #[cfg(not(target_os = "ios"))]
-    {
-        crate::infrastructure::persistence::desktop_data_dir()
-    }
-}
-
-pub fn pending_restore_dir() -> PathBuf {
-    restore_state_parent().join("pending_restore")
-}
-
-pub fn import_staging_dir() -> PathBuf {
-    restore_state_parent().join("import_staging")
-}
-
-pub fn restore_bak_dir() -> PathBuf {
-    restore_state_parent().join("restore_bak")
-}
-
-pub const RESTORE_STATE_PATH: &str = "restore_state.json";
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct RestoreState {
-    pub staged_db_crc32: u32,
-    pub floor: i64,
 }
 
 #[derive(Debug, Clone)]
